move step 'augmenting symbol information' into create graph step
